renamed package localFiles to files, as it will just handle the generic file contents
