+++ src/main/java/files/AbstractWayfindingFile.java
@@ -1,8 +1,6 @@
-package localFiles;
+package files;
 
 import com.google.api.services.drive.model.File;
-import java.io.FileInputStream;
-import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.io.InputStream;
 import nodemanager.io.DriveIOOp;
