+++ src/main/java/files/FileType.java
@@ -1,4 +1,4 @@
-package localFiles;
+package files;
 
 /**
  * Used to keep track of
