+++ src/main/java/nodemanager/gui/exportData/ExportBody.java
@@ -1,6 +1,6 @@
 package nodemanager.gui.exportData;
 
-import localFiles.WayfindingManifest;
+import files.WayfindingManifest;
 import java.awt.*;
 import java.awt.event.ItemEvent;
 import javax.swing.*;
