+++ src/main/java/nodemanager/gui/importData/DriveImportBody.java
@@ -9,9 +9,9 @@ import java.awt.Container;
 import java.awt.GridLayout;
 import java.util.ArrayList;
 import javax.swing.*;
-import localFiles.FileType;
+import files.FileType;
 import nodemanager.io.VersionLog;
-import localFiles.WayfindingManifest;
+import files.WayfindingManifest;
 
 /**
  * Acts as the body of the import dialog whenever the user clicks the import from drive button.
