+++ src/main/java/nodemanager/gui/importData/LocalImportBody.java
@@ -1,6 +1,6 @@
 package nodemanager.gui.importData;
 
-import localFiles.FileType;
+import files.FileType;
 import java.awt.Container;
 import java.awt.GridLayout;
 import java.awt.event.ActionEvent;
