+++ src/main/java/nodemanager/io/GoogleDriveUploader.java
@@ -1,6 +1,6 @@
 package nodemanager.io;
 
-import localFiles.AbstractWayfindingFile;
+import files.AbstractWayfindingFile;
 import com.google.api.client.auth.oauth2.Credential;
 import com.google.api.client.extensions.java6.auth.oauth2.AuthorizationCodeInstalledApp;
 import com.google.api.client.extensions.jetty.auth.oauth2.LocalServerReceiver;
